requirements.txt
attrs==19.3.0
aws-cdk.assets==1.36.1
aws-cdk.aws-apigateway==1.36.1
aws-cdk.aws-applicationautoscaling==1.36.1
aws-cdk.aws-autoscaling==1.36.1
aws-cdk.aws-autoscaling-common==1.36.1
aws-cdk.aws-autoscaling-hooktargets==1.36.1
aws-cdk.aws-batch==1.36.1
aws-cdk.aws-certificatemanager==1.36.1
aws-cdk.aws-cloudformation==1.36.1
aws-cdk.aws-cloudfront==1.36.1
aws-cdk.aws-cloudwatch==1.36.1
aws-cdk.aws-codebuild==1.36.1
aws-cdk.aws-codecommit==1.36.1
aws-cdk.aws-codepipeline==1.36.1
aws-cdk.aws-cognito==1.36.1
aws-cdk.aws-dynamodb==1.36.1
aws-cdk.aws-ec2==1.36.1
aws-cdk.aws-ecr==1.36.1
aws-cdk.aws-ecr-assets==1.36.1
aws-cdk.aws-ecs==1.36.1
aws-cdk.aws-elasticloadbalancing==1.36.1
aws-cdk.aws-elasticloadbalancingv2==1.36.1
aws-cdk.aws-events==1.36.1
aws-cdk.aws-events-targets==1.36.1
aws-cdk.aws-iam==1.36.1
aws-cdk.aws-kinesis==1.36.1
aws-cdk.aws-kms==1.36.1
aws-cdk.aws-lambda==1.36.1
aws-cdk.aws-lambda-event-sources==1.36.1
aws-cdk.aws-logs==1.36.1
aws-cdk.aws-route53==1.36.1
aws-cdk.aws-route53-targets==1.36.1
aws-cdk.aws-s3==1.36.1
aws-cdk.aws-s3-assets==1.36.1
aws-cdk.aws-s3-notifications==1.36.1
aws-cdk.aws-sam==1.36.1
aws-cdk.aws-secretsmanager==1.36.1
aws-cdk.aws-servicediscovery==1.36.1
aws-cdk.aws-sns==1.36.1
aws-cdk.aws-sns-subscriptions==1.36.1
aws-cdk.aws-sqs==1.36.1
aws-cdk.aws-ssm==1.36.1
aws-cdk.aws-stepfunctions==1.36.1
aws-cdk.cloud-assembly-schema==1.36.1
aws-cdk.core==1.36.1
aws-cdk.custom-resources==1.36.1
aws-cdk.cx-api==1.36.1
aws-cdk.region-info==1.36.1
cattrs==1.0.0
constructs==3.0.2
importlib-metadata==1.6.0
jsii==1.4.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
publication==0.0.3
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
six==1.14.0
typing-extensions==3.7.4.2
wcwidth==0.1.9
zipp==3.1.0
